mcq_extraction/requirements.txt
transformers==4.28.0
datasets>=2.14.6,<3.0.0
torch==2.0.0
PyMuPDF>=1.26.3,<1.27.0
python-docx==0.8.11
pytesseract==0.3.10
pillow>=11.3.0,<12.0
nltk>=3.9.1,<4.0.0
pdfplumber==0.10.2
numpy>=1.26.4,<2.0.0
pyarrow>=6.0.0,<21.0.0
seqeval==1.2.2
pyyaml>=6.0
opencv-python==4.7.0.72
label-studio>=1.17.0,<1.18.0
